scraping article meta-data
"scraping_utils.py" contains the meat of this update. Enter an URL to a Cochrane article (could be either the complete URL or the https://doi.org/xxxxx URL), it returns the number of guidelines citing that article (if any), and the publication date. If the URL leads to an older version of the article, the function identifies the notice for a latest version and fetch data from the latest version instead. More features to come.